created seach data, install vue-cleave-component and sweetalter2 VDO 07

+++ resources/js/Pages/Store.vue
@@ -18,16 +18,16 @@
           <input type="text" class="form-control mb-2" v-model="FormStore.name" placeholder="..." >
 
           <label  class="form-label fs-6">ຈຳນວນ:</label>
-          <input type="text" class="form-control mb-2" v-model="FormStore.qty"  placeholder="..." >
+          <cleave :options="options" class="form-control mb-2" v-model="FormStore.qty"  placeholder="..." />
 
           <div class="row">
             <div class="col-md-6">
                 <label  class="form-label fs-6">ລາຄາຊື້:</label>
-                    <input type="text" class="form-control mb-2" v-model="FormStore.price_buy" placeholder="..." >
+                    <cleave :options="options" class="form-control mb-2" v-model="FormStore.price_buy" placeholder="..." />
             </div>
             <div class="col-md-6">
                 <label  class="form-label fs-6">ລາຄາຂາຍ:</label>
-                    <input type="text" class="form-control mb-2" v-model="FormStore.price_sell"  placeholder="..." >
+                    <cleave :options="options" class="form-control mb-2" v-model="FormStore.price_sell"  placeholder="..." />
             </div>
           </div>
           
@@ -51,7 +51,7 @@
                 </select>
             </div>
             <div class=" d-flex">
-                <input type="text" class="form-control me-2" placeholder="ຄົ້ນຫາ...">
+                <input type="text" class="form-control me-2" v-model="Search" @keyup.enter="GetStore()" placeholder="ຄົ້ນຫາ...">
                 <button type="button" class="btn btn-primary" @click="AddStore()" >ເພີ່ມຂໍ້ມູນ</button>
             </div>
         </div> 
@@ -62,8 +62,8 @@
             <th>ID</th>
             <th>ຮູບ</th>
             <th>ຊື່ສິນຄ້າ</th>
-            <th>ລາຄາຊື້</th>
-            <th>ຈັດການ</th>
+            <th class="text-center">ລາຄາຊື້</th>
+            <th class="text-center">ຈັດການ</th>
           </tr>
         </thead>
         <tbody>
@@ -73,8 +73,8 @@
             <td>
              {{ list.name }}
             </td>
-            <td>{{ list.price_buy }}</td>
-            <td>
+            <td class="text-end">{{ formatPrice(list.price_buy) }} ກີບ</td>
+            <td class="text-center">
               <div class="dropdown">
                 <button type="button" class="btn p-0 dropdown-toggle hide-arrow" data-bs-toggle="dropdown"><i class="bx bx-dots-vertical-rounded"></i></button>
                 <div class="dropdown-menu">
@@ -88,6 +88,7 @@
           
         </tbody>
       </table>
+      <button @click="showAlert">Hello world</button>
 
       <Pagination :pagination="StoreData" :offset="4" @paginate="GetStore($event)" />
 
@@ -117,7 +118,18 @@ export default {
             FormType:true,
             EditID:'',
             Sort:'asc',
-            perpage:5
+            perpage:5,
+            Search:'',
+            options: {
+                  numeral: true,
+                  numeralPositiveOnly: true,
+                  noImmediatePrefix: true,
+                  rawValueTrimPrefix: true,
+                  numeralIntegerScale: 10,
+                  numeralDecimalScale: 2,
+                  numeralDecimalMark: ',',
+                  delimiter: '.'
+                }
         }
     },
     computed:{
@@ -130,6 +142,19 @@ export default {
         }
     },
     methods:{
+        showAlert() {
+        // Use sweetalert2
+        this.$swal({
+            title: "res.data.message",
+                                showConfirmButton: false,
+                                icon:"error",
+                                timer:3500
+            });
+        },
+        formatPrice(value) {
+            let val = (value / 1).toFixed(0).replace(",", ".");
+            return val.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ".");
+        },
         ChangeSort(){
             if(this.Sort == 'asc'){
                 this.Sort = 'desc'
@@ -161,16 +186,68 @@ export default {
 
             }).catch((error)=>{
                 console.log(error)
+                if(error.response.status == 401){
+                            this.store.remove_token()
+                            this.store.remove_user()
+                            localStorage.removeItem('web_user')
+                            localStorage.removeItem('web_token')
+                            this.$router.push('/login')
+                            console.log('clear')
+                            }
             })
         },
         DelStore(id){
-            axios.delete(`api/store/delete/${id}`,{ headers:{ Authorization:'Bearer '+ this.store.get_token }}).then((res)=>{
-                if(res.data.success){
-                    this.GetStore()
-                }
-            }).catch((error)=>{
-                console.log(error)
-            })
+
+
+            this.$swal({
+                    title: "ທ່ານແນ່ໃຈບໍ່?",
+                    text: "ທີ່ຈະລຶບຂໍ້ມູນນີ້!",
+                    icon: "warning",
+                    showCancelButton: true,
+                    confirmButtonColor: "#3085d6",
+                    cancelButtonColor: "#d33",
+                    confirmButtonText: "ຕົກລົງ!",
+                    cancelButtonText:"ຍົກເລີກ"
+                    }).then((result) => {
+                    if (result.isConfirmed) {
+
+
+                        axios.delete(`api/store/delete/${id}`,{ headers:{ Authorization:'Bearer '+ this.store.get_token }}).then((res)=>{
+                                if(res.data.success){
+                                    this.GetStore()
+                                    this.$swal({
+                                        title: "ການລຶບຂໍ້ມູນ!",
+                                        text: res.data.message,
+                                        icon: "success",
+                                        showConfirmButton: false,
+                                        timer:3500
+                                        });
+                                } else {
+                                    this.$swal({
+                                        title: res.data.message,
+                                        showConfirmButton: false,
+                                        icon:"error",
+                                        timer:3500
+                                    });
+                                }
+                            }).catch((error)=>{
+                                console.log(error)
+                                if(error.response.status == 401){
+                                            this.store.remove_token()
+                                            this.store.remove_user()
+                                            localStorage.removeItem('web_user')
+                                            localStorage.removeItem('web_token')
+                                            this.$router.push('/login')
+                                            console.log('clear')
+                                            }
+                            })
+                    
+                        
+                    }
+                    });
+
+            
+            
         },
         CancelStore(){
             this.ShowForm = false
@@ -184,11 +261,35 @@ export default {
                             this.ShowForm = false
                             this.GetStore()
 
+                            this.$swal({
+                                position:'top-end',
+                                toast:true,
+                                title: res.data.message,
+                                showConfirmButton: false,
+                                icon:"success",
+                                timer:2500
+                            });
+
                         } else {
+
+                            this.$swal({
+                                title: res.data.message,
+                                showConfirmButton: false,
+                                icon:"error",
+                                timer:3500
+                            });
                             
                         }
                     }).catch((error)=>{
                         console.log(error)
+                        if(error.response.status == 401){
+                            this.store.remove_token()
+                            this.store.remove_user()
+                            localStorage.removeItem('web_user')
+                            localStorage.removeItem('web_token')
+                            this.$router.push('/login')
+                            console.log('clear')
+                            }
                     })
 
 
@@ -200,23 +301,65 @@ export default {
                                 this.ShowForm = false
                                 this.GetStore()
 
+                                this.$swal({
+                                position:'top-end',
+                                toast:true,
+                                title: res.data.message,
+                                showConfirmButton: false,
+                                icon:"success",
+                                timer:2500
+                            });
+
+                                } else {
+                                    this.$swal({
+                                title: res.data.message,
+                                showConfirmButton: false,
+                                icon:"error",
+                                timer:3500
+                            });
                                 }
                         }).catch((error)=>{
                         console.log(error)
+                        if(error.response.status == 401){
+                            this.store.remove_token()
+                            this.store.remove_user()
+                            localStorage.removeItem('web_user')
+                            localStorage.removeItem('web_token')
+                            this.$router.push('/login')
+                            console.log('clear')
+                            }
                     })
 
                 }
         },
         GetStore(page){
-            axios.get(`api/store?page=${page}&sort=${this.Sort}&perpage=${this.perpage}`,{ headers:{ Authorization:'Bearer '+ this.store.get_token }}).then((res)=>{
+            axios.get(`api/store?page=${page}&sort=${this.Sort}&perpage=${this.perpage}&search=${this.Search}`,{ headers:{ Authorization:'Bearer '+ this.store.get_token }}).then((res)=>{
                 this.StoreData = res.data
             }).catch((error)=>{
-                console.log(error)
+                console.log(error.response.status)
+                // ກວດເບີ່ງ statis code ຖ້າເປັນ 401 ໃຫ້ໄປໜ້າ Login
+                if(error.response.status == 401){
+
+                    this.store.remove_token()
+                    this.store.remove_user()
+                    localStorage.removeItem('web_user')
+                    localStorage.removeItem('web_token')
+                    this.$router.push('/login')
+                    console.log('clear')
+                }
+
             })
         }
     },
     created(){
         this.GetStore()
+    },
+    watch:{
+        Search(){
+            if(this.Search == ''){
+                this.GetStore()
+            }
+        }
     }
 }
 </script>
